backend/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
Flask==2.0.2
Flask-Cors==3.0.10
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jwcrypto==1.0
MarkupSafe==2.0.1
mysql-connector==2.2.9
oauth2client==4.1.3
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive==1.3.1
pyee==8.2.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dotenv==0.19.2
python-ffmpeg==1.0.11
python-ffmpeg-video-streaming==0.1.14
PyYAML==6.0
requests==2.27.0
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
